Set Ruby version to 2.2.2
Rack and Rails require Ruby 2.2.2 so we need to upgrade travis to use
only Ruby 2.2.2 and set the minimum version in the gemspec.

This means we'll have to split this peripheral gem into 2 releases. One
for the 4 apps and one for the 5 apps because the changes required by
Rails 5 won't work on older version.