Use render_to_string :html when not caching layout
If a layout is using a format in it's name (e.g. layout.html.erb) then
using render_to_string :plain breaks because it's looking for the text
version of the layout (e.g. layout.text.erb).

We can fix this by using render_to_string :layout, which restores the
old behavior with render_to_string :text but there's still a bug with
non-HTML layouts.

Fixes #38.